@@ -5,6 +5,8 @@
         "php": "^7.1.3",
         "ext-ctype": "*",
         "ext-iconv": "*",
+        "doctrine/annotations": "^1.6",
+        "doctrine/doctrine-bundle": "^1.9",
         "sensio/framework-extra-bundle": "^5.2",
         "symfony/console": "*",
         "symfony/flex": "^1.1",
@@ -12,11 +14,12 @@
         "symfony/framework-bundle": "*",
         "symfony/orm-pack": "^1.0",
         "symfony/twig-bundle": "*",
+        "symfony/validator": "4.1.*",
         "symfony/yaml": "*"
     },
     "require-dev": {
         "symfony/dotenv": "*",
-        "symfony/maker-bundle": "^1.7"
+        "symfony/maker-bundle": "^1.9"
     },
     "config": {
         "preferred-install": {
