@@ -5,6 +5,8 @@ namespace App\Controller;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Request;
+use App\Entity\Naudotojai;
+use Symfony\Component\HttpFoundation\Session\Session;
 
 class LoginController extends AbstractController
 {
@@ -13,9 +15,15 @@ class LoginController extends AbstractController
      */
     public function index()
     {
-        return $this->render('login/index.html.twig', [
-            'controller_name' => 'LoginController',
-        ]);
+        $session = new Session();
+        $session->start();
+        $email = $session->get("userEmail");
+        if($email)
+            return $this->redirectToRoute("home");
+
+        $data = [];
+        $data["errors"] = [];
+        return $this->render('login/index.html.twig', $data);
     }
 
     /**
@@ -28,18 +36,25 @@ class LoginController extends AbstractController
         ->add('password')
         ->getForm();
 
+        $data = [];
+        $data["errors"] = [];
+
         $form->handleRequest($request);
 
         if($form->isSubmitted()){
             $data = $form->getData();
-            return $this->redirectToRoute('home');
+            if($this->userExists($data["email"]) && $this->isPasswordCorrect($data["email"], $data["password"])){
+                $session = new Session();
+                $session->start();
+                $session->set('userEmail', $data["email"]);
+            }else{
+                $data["errors"] = [["text" => "Duomenys neteisingi."]];
+                return $this->render('login/index.html.twig', $data);
+            }
         }else{
             return $this->redirectToRoute('login');
         }
-
-        return $this->render('login/index.html.twig', [
-            'controller_name' => 'LoginController',
-        ]);
+        return $this->redirectToRoute("home");
     }
 
     /**
@@ -47,6 +62,31 @@ class LoginController extends AbstractController
      */
     public function logout()
     {
+        $session = new Session();
+        $session->start();
+        $session->invalidate();
         return $this->redirectToRoute('home');
+    }
+
+    private function userExists($email){
+        $userByEmail = $this->getDoctrine()
+        ->getRepository(Naudotojai::class)
+        ->findOneBy(["el_pastas" => $email]);
+
+        if(!$userByEmail){
+            return false;
+        }
+
+        return true;
+    }
+
+    private function isPasswordCorrect($email, $entPass){
+        $userByEmail = $this->getDoctrine()
+        ->getRepository(Naudotojai::class)
+        ->findOneBy(["el_pastas" => $email]);
+
+        $pass = $userByEmail->getSlaptazodis();
+
+        return password_verify($entPass, $pass);
     }
 }
