@@ -51,17 +51,17 @@ class __TwigTemplate_7dd1a9b75b31e52fddbf81a2416e7f7f8a5a630e923df7af9d747b800c1
         ";
         // line 15
         $this->displayBlock('navbar', $context, $blocks);
-        // line 52
+        // line 56
         echo "
 
         <!-- End Top Nav Bar -->
         ";
-        // line 55
+        // line 59
         $this->displayBlock('body', $context, $blocks);
-        // line 62
+        // line 66
         echo "        ";
         $this->displayBlock('javascripts', $context, $blocks);
-        // line 63
+        // line 67
         echo "    </body>
 </html>
 ";
@@ -120,50 +120,63 @@ class __TwigTemplate_7dd1a9b75b31e52fddbf81a2416e7f7f8a5a630e923df7af9d747b800c1
                         </li>\t\t\t\t\t
                     </ul>
                     <ul class=\"navbar-nav\">
-                            <li class=\"nav-item\">
+                            ";
+        // line 28
+        if (twig_test_empty(twig_get_attribute($this->env, $this->source, twig_get_attribute($this->env, $this->source, (isset($context["app"]) || array_key_exists("app", $context) ? $context["app"] : (function () { throw new Twig_Error_Runtime('Variable "app" does not exist.', 28, $this->source); })()), "session", array()), "get", array(0 => "userEmail"), "method"))) {
+            // line 29
+            echo "                            <li class=\"nav-item\">
                                     <a href=\"";
-        // line 29
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("login");
-        echo "\" class=\"nav-link\">Prisijungti</a>
+            // line 30
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("login");
+            echo "\" class=\"nav-link\">Prisijungti</a>
                             </li>\t
                             <li class=\"nav-item\">
                                    <a href=\"";
-        // line 32
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("registration");
-        echo "\" class=\"nav-link\">Registruotis</a>
+            // line 33
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("registration");
+            echo "\" class=\"nav-link\">Registruotis</a>
                             </li>
-                                <li class=\"nav-item dropdown\">
+                            ";
+        }
+        // line 36
+        echo "                            ";
+        if ( !twig_test_empty(twig_get_attribute($this->env, $this->source, twig_get_attribute($this->env, $this->source, (isset($context["app"]) || array_key_exists("app", $context) ? $context["app"] : (function () { throw new Twig_Error_Runtime('Variable "app" does not exist.', 36, $this->source); })()), "session", array()), "get", array(0 => "userEmail"), "method"))) {
+            // line 37
+            echo "                                <li class=\"nav-item dropdown\">
                                     <a class=\"nav-link dropdown-toggle\" data-toggle=\"dropdown\" href=\"#\" role=\"button\" aria-haspopup=\"true\" aria-expanded=\"false\">Naudotojo Meniu</a>
                                     <div class=\"dropdown-menu\">
                                         <a class=\"dropdown-item\" href=\"";
-        // line 37
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("edit_profile");
-        echo "\">Redaguoti Profilį</a>
+            // line 40
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("edit_profile");
+            echo "\">Redaguoti Profilį</a>
                                         <a class=\"dropdown-item\" href=\"";
-        // line 38
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("user_ads");
-        echo "\">Mano skelbimai</a>
+            // line 41
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("user_ads");
+            echo "\">Mano skelbimai</a>
                                         <a class=\"dropdown-item\" href=\"";
-        // line 39
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("contracts");
-        echo "\">Mano sutartys</a>
+            // line 42
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("contracts");
+            echo "\">Mano sutartys</a>
                                         <a class=\"dropdown-item\" href=\"";
-        // line 40
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("new_add");
-        echo "\">Naujas skelbimas</a>
+            // line 43
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("new_add");
+            echo "\">Naujas skelbimas</a>
                                         <a class=\"dropdown-item\" href=\"";
-        // line 41
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("admin_panel");
-        echo "\">Administratoriaus panele</a>
+            // line 44
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("admin_panel");
+            echo "\">Administratoriaus panele</a>
                                     </div>
                                 </li>
                                 <li class=\"nav-item\">
                                     <a href=\"";
-        // line 45
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("logout");
-        echo "\" class=\"nav-link\">Atsijungti</a>
+            // line 48
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("logout");
+            echo "\" class=\"nav-link\">Atsijungti</a>
                                 </li>
-                        </ul>   
+                                ";
+        }
+        // line 51
+        echo "                        </ul>   
                 </div>
         </nav>
         <br>
@@ -173,18 +186,18 @@ class __TwigTemplate_7dd1a9b75b31e52fddbf81a2416e7f7f8a5a630e923df7af9d747b800c1
 
     }
 
-    // line 55
+    // line 59
     public function block_body($context, array $blocks = array())
     {
         $__internal_319393461309892924ff6e74d6d6e64287df64b63545b994e100d4ab223aed02 = $this->extensions["Symfony\\Bridge\\Twig\\Extension\\ProfilerExtension"];
         $__internal_319393461309892924ff6e74d6d6e64287df64b63545b994e100d4ab223aed02->enter($__internal_319393461309892924ff6e74d6d6e64287df64b63545b994e100d4ab223aed02_prof = new Twig_Profiler_Profile($this->getTemplateName(), "block", "body"));
 
-        // line 56
+        // line 60
         echo "
         ";
-        // line 57
+        // line 61
         $this->displayBlock('ads', $context, $blocks);
-        // line 60
+        // line 64
         echo "
         ";
         
@@ -192,13 +205,13 @@ class __TwigTemplate_7dd1a9b75b31e52fddbf81a2416e7f7f8a5a630e923df7af9d747b800c1
 
     }
 
-    // line 57
+    // line 61
     public function block_ads($context, array $blocks = array())
     {
         $__internal_319393461309892924ff6e74d6d6e64287df64b63545b994e100d4ab223aed02 = $this->extensions["Symfony\\Bridge\\Twig\\Extension\\ProfilerExtension"];
         $__internal_319393461309892924ff6e74d6d6e64287df64b63545b994e100d4ab223aed02->enter($__internal_319393461309892924ff6e74d6d6e64287df64b63545b994e100d4ab223aed02_prof = new Twig_Profiler_Profile($this->getTemplateName(), "block", "ads"));
 
-        // line 58
+        // line 62
         echo "
         ";
         
@@ -206,7 +219,7 @@ class __TwigTemplate_7dd1a9b75b31e52fddbf81a2416e7f7f8a5a630e923df7af9d747b800c1
 
     }
 
-    // line 62
+    // line 66
     public function block_javascripts($context, array $blocks = array())
     {
         $__internal_319393461309892924ff6e74d6d6e64287df64b63545b994e100d4ab223aed02 = $this->extensions["Symfony\\Bridge\\Twig\\Extension\\ProfilerExtension"];
@@ -224,7 +237,7 @@ class __TwigTemplate_7dd1a9b75b31e52fddbf81a2416e7f7f8a5a630e923df7af9d747b800c1
 
     public function getDebugInfo()
     {
-        return array (  210 => 62,  202 => 58,  196 => 57,  188 => 60,  186 => 57,  183 => 56,  177 => 55,  163 => 45,  156 => 41,  152 => 40,  148 => 39,  144 => 38,  140 => 37,  132 => 32,  126 => 29,  118 => 24,  108 => 17,  105 => 16,  99 => 15,  92 => 11,  86 => 10,  74 => 5,  65 => 63,  62 => 62,  60 => 55,  55 => 52,  53 => 15,  48 => 12,  46 => 10,  38 => 5,  32 => 1,);
+        return array (  223 => 66,  215 => 62,  209 => 61,  201 => 64,  199 => 61,  196 => 60,  190 => 59,  179 => 51,  173 => 48,  166 => 44,  162 => 43,  158 => 42,  154 => 41,  150 => 40,  145 => 37,  142 => 36,  136 => 33,  130 => 30,  127 => 29,  125 => 28,  118 => 24,  108 => 17,  105 => 16,  99 => 15,  92 => 11,  86 => 10,  74 => 5,  65 => 67,  62 => 66,  60 => 59,  55 => 56,  53 => 15,  48 => 12,  46 => 10,  38 => 5,  32 => 1,);
     }
 
     public function getSourceContext()
@@ -256,12 +269,15 @@ class __TwigTemplate_7dd1a9b75b31e52fddbf81a2416e7f7f8a5a630e923df7af9d747b800c1
                         </li>\t\t\t\t\t
                     </ul>
                     <ul class=\"navbar-nav\">
+                            {% if app.session.get('userEmail') is empty %}
                             <li class=\"nav-item\">
                                     <a href=\"{{ path(\"login\") }}\" class=\"nav-link\">Prisijungti</a>
                             </li>\t
                             <li class=\"nav-item\">
                                    <a href=\"{{ path(\"registration\") }}\" class=\"nav-link\">Registruotis</a>
                             </li>
+                            {% endif %}
+                            {% if app.session.get('userEmail') is not empty %}
                                 <li class=\"nav-item dropdown\">
                                     <a class=\"nav-link dropdown-toggle\" data-toggle=\"dropdown\" href=\"#\" role=\"button\" aria-haspopup=\"true\" aria-expanded=\"false\">Naudotojo Meniu</a>
                                     <div class=\"dropdown-menu\">
@@ -275,6 +291,7 @@ class __TwigTemplate_7dd1a9b75b31e52fddbf81a2416e7f7f8a5a630e923df7af9d747b800c1
                                 <li class=\"nav-item\">
                                     <a href=\"{{ path(\"logout\") }}\" class=\"nav-link\">Atsijungti</a>
                                 </li>
+                                {% endif %}
                         </ul>   
                 </div>
         </nav>
